LinkList와 LinkItem 분리
